+++ changelog.rst
@@ -6,6 +6,7 @@ Last change: 20-JAN-2025 MTS
 0.7.5
 -----
 - Plugin docs update
+- Bug fixes
 
 0.7.1 - 0.7.4
 -------------