+++ changelog.rst
@@ -9,6 +9,7 @@ Last change: 06-FEB-2025 MTS
 - Filter histogram display fixed for datasets with low variance (bug fix)
 - AIM undrifting works now if the first frames of localizations are filtered out (bug fix)
 - 2D drift plot in Render inverts y axis to match the rendered localizations
+- 3D animation fixed
 - Other minor bug fixes
 
 0.7.1 - 0.7.4